Improve verse URL
It is better to describe the current URL of a verses page (John/1/1)
that specifies a verse to be the same as the actual page that is being
rendered. Before, it was using the URL of the chapter without the verse.
This change is especially helpful when describing the metadata for open
graph, etc.