Upgrade to actix-web 1.0 (#28)
This commit upgrades actix-web to version 1.0. This required some
refactoring in order to adapt to the recommended patterns with this
version. The biggest change was to remove the explict sync arbiter
abstractions, and replace them with `web::block`, whose inputted
function calls the `db::sword_drill` (the database layer) explicitly.

Other changes include:
- Upgrade libsqlite3-sys
- Upgrade diesel
- Add Docker section to README