Improve breadcrumb JSON-LD
The schema for "ListItem" indicates that there should be an "item"
property for underneath it that is also structured data (it's currently
a string). This commit structures the data for the "item" property to a
superset of "Thing", depending on context.

Reference: https://schema.org/ListItem
Google's Reference: https://developers.google.com/search/docs/data-types/breadcrumb