Add error alerting via Sentry (#41)
This commit fixes a bug where a search value that starts with a number
and ends with a space would throw a database error which was unhandled
by the middleware. To fix this error, the query string is trimmed and an
empty result set is returned. The error handling was also improved to
log the underlying database error instead of a generic message.

In the process, I realized that it would be nice to be alerted to these
types of issues instead of accidentally noticing them in a log. Sentry
is a relatively simple option, so I added it. Unfortunately, though
Sentry provides middleware for actix-web, it doesn't work yet for
versions > 1.0. I copied and pasted code from [this
issue](https://github.com/getsentry/sentry-rust/issues/143) that lets
the middleware work in a less-featured-than-ideal manner. I also added
the Sentry panic handler.