Make sentry DSN a variable rather than hardcoded
Though it was convenient to have the Sentry (issue alerting) DSN
hardcoded into `web::main`, it was too easy for alerts to be sent
unintentionally (e.g. if the repo was forked and tested by another).
This commit makes the DSN an environment variable, which is defaulted to
`None`; thus disabling alerting if the var isn't set.