Add swipe navigation
Swipe navigation is provided by a simple touch event handler that uses
`link` elements that may be in the `head` of the document. If the
`rel="prev"` and `rel="next"` elements exist, then they are used to
allow for gesture navigation by swiping left and right on a
touch-enabled device.

Also invalidate the PWA asset cache so the newer version of the JS is
delivered.